+++ gitlint/hooks.py
@@ -22,7 +22,7 @@ class GitHookInstaller(object):
         # assert that current directory is a git repository
         hooks_dir = os.path.abspath(os.path.join(lint_config.target, HOOKS_DIR_PATH))
         if not os.path.isdir(hooks_dir):
-            raise GitHookInstallerError("{} is not a git repository.".format(lint_config.target))
+            raise GitHookInstallerError("{0} is not a git repository.".format(lint_config.target))
 
     @staticmethod
     def install_commit_msg_hook(lint_config):
@@ -30,7 +30,7 @@ class GitHookInstaller(object):
         dest_path = GitHookInstaller.commit_msg_hook_path(lint_config)
         if os.path.exists(dest_path):
             raise GitHookInstallerError(
-                "There is already a commit-msg hook file present in {}.\n".format(dest_path) +
+                "There is already a commit-msg hook file present in {0}.\n".format(dest_path) +
                 "gitlint currently does not support appending to an existing commit-msg file.")
 
         # copy hook file
@@ -44,12 +44,12 @@ class GitHookInstaller(object):
         GitHookInstaller._assert_git_repo(lint_config)
         dest_path = GitHookInstaller.commit_msg_hook_path(lint_config)
         if not os.path.exists(dest_path):
-            raise GitHookInstallerError("There is no commit-msg hook present in {}.".format(dest_path))
+            raise GitHookInstallerError("There is no commit-msg hook present in {0}.".format(dest_path))
 
         with open(dest_path) as file:
             lines = file.readlines()
             if len(lines) < 2 or lines[1] != GITLINT_HOOK_IDENTIFIER:
-                msg = "The commit-msg hook in {} was not installed by gitlint (or it was modified).\n" + \
+                msg = "The commit-msg hook in {0} was not installed by gitlint (or it was modified).\n" + \
                       "Uninstallation of 3th party or modified gitlint hooks is not supported."
                 raise GitHookInstallerError(msg.format(dest_path))