+++ gitlint/tests/test_title_rules.py
@@ -67,7 +67,7 @@ class TitleRuleTests(BaseTestCase):
         for char in punctuation:
             line = "This is a test" + char
             gitcontext = self.gitcontext(line)
-            expected_violation = RuleViolation("T3", "Title has trailing punctuation ({})".format(char), line)
+            expected_violation = RuleViolation("T3", "Title has trailing punctuation ({0})".format(char), line)
             violations = rule.validate(line, gitcontext)
             self.assertListEqual(violations, [expected_violation])